% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coale_projection.R
\name{coale_projection}
\alias{coale_projection}
\title{Coale Scenario Projection}
\usage{
coale_projection(Lx_zero, Fx_zero, k1 = -0.015, k2 = 0, rho = 0, e0_start = 80)
}
\arguments{
\item{Lx_zero}{vector of initial Lx values}

\item{Fx_zero}{vector if initial Fx values}

\item{k1}{rate of NRR decline}

\item{k2}{additional curvature of NRR change}

\item{rho}{rate of e0 change}

\item{e0_start}{initial e0 value}
}
\description{
Specialized projection over 100 years, where NRR declines
exponentially.
}
\details{
\eqn{exp(k1 * t + k2 * t^2)}

should do the math formula for exp(exp(k1....)
}
