% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japan_asfr.R
\docType{data}
\name{japan_asfr}
\alias{japan_asfr}
\title{japan_asfr}
\format{
A data frame
\describe{
\item{Year}{year}
\item{Age}{age group}
\item{ASFR}{age specific fertility rate}
}
}
\source{
Human Fertility Database. Max Planck Institute for Demographic Research (Germany)
and Vienna Institute of Demography (Austria).
Available at www.humanfertility.org (data downloaded on July 28th, 2022).
}
\usage{
japan_asfr
}
\description{
A data set containing age specific fertility rates
from Japan. Data downloaded from HFD July 2022.
}
\keyword{datasets}
