% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leslie1.R
\name{leslie1}
\alias{leslie1}
\title{Leslie Matrix with 1-Year Age Groups}
\usage{
leslie1(Lx, Fx)
}
\arguments{
\item{Lx}{vector of 1-year Lx values}

\item{Fx}{vector of 1-year Fx values}
}
\value{
A Leslie metrix
}
\description{
Makes a Leslie matrix out of 1-year age-interval Lx and Fx schedules
}
