% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifetable_to_matrix.R
\name{lifetable_to_matrix}
\alias{lifetable_to_matrix}
\title{Convert lifetable to matrix}
\usage{
lifetable_to_matrix(data, years, rate = c("Lx", "ASFR"), scaler = 1e+05)
}
\arguments{
\item{data}{data from HMD/HFD}

\item{years}{a range of years to select rates from}

\item{rate}{which vital rate to select, either ASFR or Lx}

\item{scaler}{radix}
}
\value{
Returns a list with the following components
}
\description{
Takes HMD/HFD structured data and turns it into a matrix
for use in project_leslie1_dynamic(). Data need to have
Year, Age, and either and Lx or ASFR column.
}
