% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_asfr.R
\docType{data}
\name{norway_asfr}
\alias{norway_asfr}
\title{norway_asfr}
\format{
A data frame
\describe{
\item{Year}{year}
\item{Age}{age group}
\item{ASFR}{age specific fertility rate}
}
}
\source{
Human Fertility Database. Max Planck Institute for Demographic Research (Germany)
and Vienna Institute of Demography (Austria).
Available at www.humanfertility.org (data downloaded on July 28th, 2022).
}
\usage{
norway_asfr
}
\description{
A data set containing age specific fertility rates
from Norway. Data downloaded from HFD July 2022.
}
\keyword{datasets}
