% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_female_lifetable.R
\docType{data}
\name{norway_female_lifetable}
\alias{norway_female_lifetable}
\title{Norway Female Lifetable}
\format{
data frame
}
\source{
HMD. Human Mortality Database. Max Planck Institute for Demographic Research (Germany),
University of California, Berkeley (USA), and French Institute for
Demographic Studies (France). Available at www.mortality.org
(data downloaded on July 28th, 2022).
}
\usage{
norway_female_lifetable
}
\description{
From HMD July 2022.
}
\keyword{datasets}
