% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway_population_size.R
\docType{data}
\name{norway_population_size}
\alias{norway_population_size}
\title{norway_population_size}
\format{
A data frame
\describe{
\item{Year}{year}
\item{Age}{age group}
\item{Female}{female population}
\item{Male}{male population}
\item{Total}{total population}
}
}
\source{
Human Fertility Database. Max Planck Institute for Demographic Research (Germany)
and Vienna Institute of Demography (Austria).
Available at www.humanfertility.org (data downloaded on July 28th, 2022).
}
\usage{
norway_population_size
}
\description{
A data set containing Norway population by age groups.
Data downloaded from HFD July 2022.
}
\keyword{datasets}
