% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_fx_vector.R
\name{pad_fx_vector}
\alias{pad_fx_vector}
\title{Pad Fx}
\usage{
pad_fx_vector(fx)
}
\arguments{
\item{fx}{a vector of ASFRs for ages 12-55}
}
\value{
a vector length 111
}
\description{
ASFRs from HMD are measured at age 12-55. This method
Pads a vector of ASFRs with zeros for ages 0-11 and 55-110,
so that it contains the same age groups as a lifetable from HMD
}
