% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_projection.R
\name{plot_projection}
\alias{plot_projection}
\title{Plot Projection}
\usage{
plot_projection(
  projection_object,
  include = c("population", "births", "nrr", "e0"),
  plot_predicted_peaks = FALSE,
  plot_generational_slopes = FALSE,
  generation_length = 30,
  generation_t = 50,
  plot_replacement_nrr = TRUE
)
}
\arguments{
\item{projection_object}{output from a projection method}

\item{include}{which plots to include, not yet implemented}

\item{plot_predicted_peaks}{calculate and mark the times of peak
population and peak births}

\item{plot_generational_slopes}{plot line segments showing average change in
population and births for one generation}

\item{generation_length}{length of one generation in years, defaults to 30}

\item{generation_t}{time at which to plot generational slope. Slope will span
between (generation_t - generation_length, generation_t)}

\item{plot_replacement_nrr}{mark replacement fertility (NRR=1) on the NRR plot}
}
\description{
Plot the results of running one of the projection methods. By default
will plot population, births, NRR, and life expectancy at birth over
time in a multi-panel plot
}
