% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_leslie1.R
\name{project_leslie1}
\alias{project_leslie1}
\title{Project with 1-Year Leslie Matrix}
\usage{
project_leslie1(
  nsteps,
  nrr.vec,
  e0.vec,
  Lx.zero,
  Fx.zero,
  starting.pop = NULL,
  normalize = TRUE,
  time = NULL
)
}
\arguments{
\item{nsteps}{number of steps (years) to project for}

\item{nrr.vec}{vector of net reproductive rates at each step in the projection.
must be length(nsteps)}

\item{e0.vec}{vector of life expectancies at each step in the projection. must
be length(nsteps)}

\item{Lx.zero}{vector of initial Lx values}

\item{Fx.zero}{vector of initial Fx values}

\item{starting.pop}{optional vector of initial population by 1-year age groups. If left unspecified,
projection will start from the stable age structure of the initial Leslie matrix}

\item{normalize}{whether to normalize the population or something IDK this isn't implemented yet}
}
\value{
A Matrix
}
\description{
Project a given population for n years with
changing leslie matrices
}
\details{
This is where we explain how this is calculated
}
